requirements.txt
mediapipe==0.10.14
opencv-python>=4.8.0
transformers>=4.40.0
torch>=2.1.0
python-rtmidi>=1.5.8
mido>=1.3.0
pygame>=2.5.0
soundfile>=0.12.1
pydub>=0.25.1
numpy>=1.24.0
PyQt5>=5.15.0
reportlab>=4.0.0
librosa>=0.10.0
matplotlib>=3.7.0
Pillow>=10.0.0
psutil>=5.9.0
ffmpeg-python>=0.2.0
